Reorganize enum classes
Hopefully this leaves the smlauncher package less cluttered
